requirements-edge.txt
opencv-python-headless==4.13.0.90
numpy==2.4.2
paho-mqtt==2.1.0
